feat: support zsh completion
For zim:fw, which is a zsh package manager
