+++ src/commands/nr.ts
@@ -1,35 +1,13 @@
 import type { Choice } from '@posva/prompts'
-import type { RunnerContext } from '../runner'
 import process from 'node:process'
 import prompts from '@posva/prompts'
 import { byLengthAsc, Fzf } from 'fzf'
-import { rawCompletionScript } from '../completion'
-import { getPackageJSON } from '../fs'
+import { getCompletionSuggestions, rawBashCompletionScript, rawZshCompletionScript } from '../completion'
+import { readPackageScripts } from '../package'
 import { parseNr } from '../parse'
 import { runCli } from '../runner'
 import { dump, load } from '../storage'
 import { limitText } from '../utils'
-
-function readPackageScripts(ctx: RunnerContext | undefined) {
-  // support https://www.npmjs.com/package/npm-scripts-info conventions
-  const pkg = getPackageJSON(ctx)
-  const rawScripts = pkg.scripts || {}
-  const scriptsInfo = pkg['scripts-info'] || {}
-
-  const scripts = Object.entries(rawScripts)
-    .filter(i => !i[0].startsWith('?'))
-    .map(([key, cmd]) => ({
-      key,
-      cmd,
-      description: scriptsInfo[key] || rawScripts[`?${key}`] || cmd,
-    }))
-
-  if (scripts.length === 0 && !ctx?.programmatic) {
-    console.warn('No scripts found in package.json')
-  }
-
-  return scripts
-}
 
 runCli(async (agent, args, ctx) => {
   const storage = await load()
@@ -39,29 +17,39 @@ runCli(async (agent, args, ctx) => {
   if (args[0] === '--completion') {
     const compLine = process.env.COMP_LINE
     const rawCompCword = process.env.COMP_CWORD
+    // In bash
     if (compLine !== undefined && rawCompCword !== undefined) {
       const compCword = Number.parseInt(rawCompCword, 10)
       const compWords = args.slice(1)
       // Only complete the second word (nr __here__ ...)
       if (compCword === 1) {
-        const raw = readPackageScripts(ctx)
-        const fzf = new Fzf(raw, {
-          selector: item => item.key,
-          casing: 'case-insensitive',
-          tiebreakers: [byLengthAsc],
-        })
-
-        // compWords will be ['nr'] when the user does not type anything after `nr` so fallback to empty string
-        const results = fzf.find(compWords[1] || '')
+        const suggestions = getCompletionSuggestions(compWords, ctx)
 
         // eslint-disable-next-line no-console
-        console.log(results.map(r => r.item.key).join('\n'))
+        console.log(suggestions.join('\n'))
       }
     }
+    // In other shells, return suggestions directly
     else {
+      const suggestions = getCompletionSuggestions(args, ctx)
+
       // eslint-disable-next-line no-console
-      console.log(rawCompletionScript)
+      console.log(suggestions.join('\n'))
     }
+    return
+  }
+
+  // Print ZSH completion script
+  if (args[0] === '--completion-zsh') {
+    // eslint-disable-next-line no-console
+    console.log(rawZshCompletionScript)
+    return
+  }
+
+  // Print Bash completion script
+  if (args[0] === '--completion-bash') {
+    // eslint-disable-next-line no-console
+    console.log(rawBashCompletionScript)
     return
   }
