+++ src/completion.ts
@@ -1,5 +1,9 @@
+import type { RunnerContext } from '.'
+import { byLengthAsc, Fzf } from 'fzf'
+import { readPackageScripts } from './package'
+
 // Print completion script
-export const rawCompletionScript = `
+export const rawBashCompletionScript = `
 ###-begin-nr-completion-###
 
 if type complete &>/dev/null; then
@@ -16,3 +20,29 @@ fi
 
 ###-end-nr-completion-###
 `.trim()
+
+export const rawZshCompletionScript = `
+#compdef nr
+
+_nr_completion() {
+  local -a completions
+  completions=("\${(f)$(nr --completion $words[2,-1])}")
+  
+  compadd -a completions
+}
+
+_nr_completion
+`.trim()
+
+export function getCompletionSuggestions(args: string[], ctx: RunnerContext | undefined) {
+  const raw = readPackageScripts(ctx)
+  const fzf = new Fzf(raw, {
+    selector: item => item.key,
+    casing: 'case-insensitive',
+    tiebreakers: [byLengthAsc],
+  })
+
+  const results = fzf.find(args[1] || '')
+
+  return results.map(r => r.item.key)
+}
